fix: solved issue with get initiatives data endpoint
