+++ src/Application/Services/Initiatives/InitiativeService.cs
@@ -98,6 +98,25 @@ public class InitiativeService : ServiceRead<Initiative, InitiativeDto, int, Ini
     }
 
     /// <summary>
+    /// Get entities by user name.
+    /// </summary>
+    /// <param name="userName">Initiative identifier.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Process result.</returns>
+    public async Task<CustomWebResponse> GetByUserNameAsync(string userName, CancellationToken ct = default)
+    {
+        var dataListEntity = await entityRepository.ListAsync(InitiativeSpec.UserNameSpec(userName), ct);
+
+        var dataListDto = dataListEntity
+            .Select(mapper.Map);
+
+        return new()
+        {
+            ResponseBody = dataListDto,
+        };
+    }
+
+    /// <summary>
     /// Get entity polygon.
     /// </summary>
     /// <param name="id">Element identifier.</param>
