+++ src/WebApi/Controllers/Rest/Auth/AuthController.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 using IAVH.BioTablero.CM.Application.Interfaces.Services;
 using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples;
-using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.InitiativeUser;
+using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Auth;
 using IAVH.BioTablero.CM.WebApi.Interfaces;
 using IAVH.BioTablero.CM.WebApi.Utils;
 
@@ -19,13 +19,13 @@ using Swashbuckle.AspNetCore.Filters;
 /// Authentication data controller.
 /// </summary>
 /// <param name="webTools">General web tools.</param>
-/// <param name="initiativeUserService">Entity service.</param>
+/// <param name="initiativeService">Inititive service.</param>
 [ApiController]
 [Route("[controller]")]
 [Produces("application/json")]
 [ApiConventionType(typeof(CustomApiConventions))]
 public class AuthController(IWebTools webTools,
-    IInitiativeUserService initiativeUserService) : ControllerBase
+    IInitiativeService initiativeService) : ControllerBase
 {
     /// <summary>
     /// Get inititives data from authenticated user.
@@ -34,10 +34,10 @@ public class AuthController(IWebTools webTools,
     /// <returns>Selected entity data.</returns>
     [Authorize]
     [HttpGet("InitiativesData")]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeUserListResponseExample))]
+    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeLoggedUserDataResponseExample))]
     public async Task<IActionResult> GetInitiativesData(CancellationToken ct)
     {
-        var response = await initiativeUserService.GetByUserNameAsync(HttpContext.GetUserName(), ct);
+        var response = await initiativeService.GetByUserNameAsync(HttpContext.GetUserName(), ct);
         return webTools.CustomResponse(response);
     }
 }
